Read the description
So I moved all of the sprite sheets out of the resources folder and
moved them into different subfolders of the project for organization,
while doing so, they reverted to textures. Sigh. So I created an
animations folder for the animations with the same sub folder setup so
that it is easier to look at in the heiarchy. Ideally I will be able to
drag and drop these folders into the resources folder when I am ready to
expand another generation.

TL;DR I took everything out of the resrouce folder and put them in
subfolders for organization and lost a bunch of texture changes